requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
werkzeug
openpyxl
qrcode[pil]
pillow
supabase
